+++ src/main/java/com/vi/counselingtoolsservice/api/service/budibase/BudibaseProxyService.java
@@ -1,27 +1,25 @@
 package com.vi.counselingtoolsservice.api.service.budibase;
 
-import com.vi.counselingtoolsservice.budibaseApi.generated.ApiClient;
-import com.vi.counselingtoolsservice.budibaseApi.generated.web.DefaultApi;
-import com.vi.counselingtoolsservice.budibaseApi.generated.web.model.App;
-import com.vi.counselingtoolsservice.budibaseApi.generated.web.model.AssignToolsRequest;
-import com.vi.counselingtoolsservice.budibaseApi.generated.web.model.User;
 import com.vi.counselingtoolsservice.config.CacheManagerConfig;
 import java.net.URI;
 import java.net.URISyntaxException;
-import java.util.ArrayList;
-import java.util.HashMap;
+import java.util.Arrays;
+import java.util.Base64;
+import java.util.HashSet;
 import java.util.Iterator;
-import java.util.LinkedHashMap;
 import java.util.List;
-import java.util.stream.Collectors;
+import java.util.Map;
+import java.util.Optional;
+import java.util.Set;
+import javax.servlet.http.HttpServletRequest;
+import javax.ws.rs.BadRequestException;
+import javax.ws.rs.NotAllowedException;
+import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.json.JSONArray;
 import org.json.JSONObject;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.cache.annotation.Cacheable;
-import org.springframework.context.annotation.Bean;
 import org.springframework.http.HttpEntity;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpMethod;
@@ -32,6 +30,7 @@ import org.springframework.web.util.UriComponentsBuilder;
 
 @Service
 @Slf4j
+@RequiredArgsConstructor
 public class BudibaseProxyService {
 
   @Value("${budibase.proxy.host}")
@@ -39,6 +38,48 @@ public class BudibaseProxyService {
 
   @Value("${budibase.proxy.port}")
   private Integer proxyServicePort;
+
+  @Value("${budibase.proxy.whitelisted}")
+  private String whitelistedURIs;
+
+  private final BudibaseApiService budibaseApiService;
+
+  public boolean isWhiteListed(HttpServletRequest request) {
+    Set<String> whitelistedURIs = new HashSet<>(Arrays.asList(this.whitelistedURIs.split(";")));
+    return whitelistedURIs.contains(request.getRequestURI());
+  }
+
+  public String extractUserIdFromJWT(HttpServletRequest request) {
+    Base64.Decoder decoder = Base64.getUrlDecoder();
+
+    String[] cookies = request.getHeader("cookie").split(";");
+
+    Optional<String> authCookie = Arrays.stream(cookies)
+        .filter(cookie -> cookie.contains("budibase:auth") && !cookie.equals("budibase:auth="))
+        .findFirst();
+
+    String token = authCookie.get().split("=")[1];
+    String[] chunks = token.split("\\.");
+    String payload = new String(decoder.decode(chunks[1]));
+    JSONObject jsonObject = new JSONObject(payload);
+    String budibaseUserId = (String) jsonObject.get("userId");
+    return budibaseUserId.substring(3);
+  }
+
+  public void validateConsultantRequest(String body, HttpMethod method,
+      HttpServletRequest request) {
+    String consultantId = extractUserIdFromJWT(request);
+    List<String> consultantAssignedUsers = budibaseApiService
+        .getConsultantAssignedUsers(consultantId);
+    String userId = extractUserIdFromBodyReadOperation(method, body);
+    Optional<String> matchedUserId = consultantAssignedUsers.stream()
+        .filter(el -> el.equals(userId)).findFirst();
+    if (matchedUserId.isEmpty()) {
+      throw new NotAllowedException(
+          "You don't have permissions to access user specific data for user with id: " + userId
+      );
+    }
+  }
 
 
   @Cacheable(cacheNames = CacheManagerConfig.TOKEN_CACHE)
@@ -98,4 +139,48 @@ public class BudibaseProxyService {
       return false;
     }
   }
+
+  private String extractUserIdFromBodyReadOperation(HttpMethod method, String body) {
+
+    if (method.equals(HttpMethod.GET)) {
+      throw new IllegalStateException();
+    }
+    JSONObject bodyJSONObject = new JSONObject(body);
+    JSONObject query = (JSONObject) bodyJSONObject.get("query");
+    JSONObject equalOperation = (JSONObject) query.get("equal");
+
+    Map<String, Object> filters = equalOperation.toMap();
+    Optional<String> filterName = filters.keySet().stream().filter(el -> el.contains("user_id"))
+        .findFirst();
+
+    if (filterName.isEmpty()) {
+      throw new BadRequestException(
+          "You should not be able to reach this point, hence in case it's a datasource without user_id than it should be filtered out before");
+    }
+
+    return (String) filters.get(filterName.get());
+  }
+
+  public void validateUserRequest(String body, HttpMethod method, HttpServletRequest request) {
+    String userIdJWT = extractUserIdFromJWT(request);
+    String userIdBody;
+    if (HttpMethod.POST.equals(method) && request.getRequestURI().contains("rows")) {
+      userIdBody = extractUserIdFromBodyUpdateOperation(body);
+    } else {
+      userIdBody = extractUserIdFromBodyReadOperation(method, body);
+    }
+
+    if (!userIdJWT.equals(userIdBody)) {
+      throw new NotAllowedException(
+          "You are not allowed to access data for user_id " + userIdBody);
+    }
+
+  }
+
+  private String extractUserIdFromBodyUpdateOperation(String body) {
+    JSONObject bodyJSONObject = new JSONObject(body);
+    return (String) bodyJSONObject.get("bb_user_id");
+  }
+
+
 }
