+++ src/main/resources/application.properties
@@ -23,6 +23,10 @@ budibase.api.url=
 budibase.api.key=
 budibase.appsApp.id=
 budibase.apps.query.id=
+budibase.proxy.whitelisted=
+budibase.public.url=
+budibase.proxy.host=proxy-service.budibase
+budibase.proxy.port=10000
 
 spring.jpa.open-in-view=true
 spring.jpa.hibernate.ddl-auto=none
